requirements.txt
langchain
pandas
PyPDF2
openai
streamlit
python-dotenv
langchain-openai
langchain-community

-e .